pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aicog-v2"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A production-ready AI Gateway SDK with multi-provider support, Redis caching, and persistent request logging."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.24.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "groq>=0.4.0",
    "openai>=1.0.0",
    "anthropic>=0.5.0",
]

[project.urls]
"Homepage" = "https://github.com/your-repo/aicog-v2"
"Bug Tracker" = "https://github.com/your-repo/aicog-v2/issues"
